+++ src/widgets/TerminalTab.vala
@@ -151,7 +151,7 @@ public class Terminal.TerminalTab : Gtk.Box {
     menu.append_section (null, preferences_section);
 
     bottom_section.append (_("Keyboard Shortcuts"), "win.show-help-overlay");
-    bottom_section.append (_("About"), "app.about");
+    bottom_section.append (_("About Black Box"), "app.about");
     menu.append_section (null, bottom_section);
 
     var pop = new Gtk.PopoverMenu.from_model (menu) {